AIHotel/pyproject.toml
[project]
name = "aihotel"
version = "0.1.0"
description = "Intelligent Hotel Recommendation Agent using LangGraph and Agentic RAG"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-groq>=0.0.1",
    "langchain-chroma>=0.1.0",
    "langchain-huggingface>=0.0.1",
    "langgraph>=0.0.20",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "streamlit>=1.30.0",
    "asyncpg>=0.29.0",
    "nest-asyncio>=1.6.0",
    "aiosqlite>=0.22.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "tools", "agents", "config"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

